Removed maven modules for unneeded example clients. Examples still failing tests.
