+++ example/mapreduce-example/mapreduce-example-runner/pom.xml
@@ -41,7 +41,7 @@
         </dependency>
         <dependency>
             <groupId>${project.parent.groupId}</groupId>
-            <artifactId>mapreduce-example-client</artifactId>
+            <artifactId>example-model</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
