+++ example/mapreduce-example/mapreduce-example-runner/src/main/resources/example/exampleObj_file1.txt
@@ -1,0 +1,8 @@
+item1a,public,1
+item1b,public,10
+item1c,public,20
+item1d,private,20
+item2a,public,1
+item2b,public,10
+item2c,public,20
+item2d,private,20
