+++ example/multi-jvm-example/multi-jvm-example-runner/src/main/java/uk/gov/gchq/palisade/example/MultiJvmExample.java
@@ -20,7 +20,8 @@ import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import uk.gov.gchq.palisade.example.client.ExampleSimpleRestClient;
+import uk.gov.gchq.palisade.client.SimpleRestServices;
+import uk.gov.gchq.palisade.example.client.ExampleSimpleClient;
 
 import java.io.File;
 import java.io.IOException;
@@ -38,7 +39,7 @@ public class MultiJvmExample {
     }
 
     public void run() throws Exception {
-        final ExampleSimpleRestClient client = new ExampleSimpleRestClient(FILE);
+        final ExampleSimpleClient client = new ExampleSimpleClient(new SimpleRestServices(), FILE);
 
         LOGGER.info("");
         LOGGER.info("Alice is reading file1...");
