+++ example/multi-jvm-example/pom.xml
@@ -33,7 +33,6 @@
         <module>multi-jvm-example-rest-resource-service</module>
         <module>multi-jvm-example-rest-user-service</module>
         <module>multi-jvm-example-rest-data-service</module>
-        <module>multi-jvm-example-client</module>
         <module>multi-jvm-example-docker-services</module>
         <module>multi-jvm-example-runner</module>
     </modules>
